# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "FyAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/refresh": {
            "post": {
                "description": "Route to generate new access token",
                "tags": [
                    "auth"
                ],
                "summary": "Refresh token",
                "responses": {
                    "201": {
                        "description": "Successfully regenerate token",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    },
                    "401": {
                        "description": "Failed to regenerate token",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/login/google": {
            "get": {
                "description": "Oauth2 Google login entry",
                "tags": [
                    "login"
                ],
                "summary": "Google OAuth2 Login",
                "parameters": [
                    {
                        "type": "string",
                        "default": "/dash",
                        "description": "Redirect URL after login",
                        "name": "redirect",
                        "in": "query"
                    }
                ],
                "responses": {
                    "303": {
                        "description": "Redirected to Google login page",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Failed to login",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/login/google/callback": {
            "get": {
                "description": "Route to handle Google callback",
                "tags": [
                    "login"
                ],
                "summary": "Google OAuth2 Callback",
                "responses": {
                    "303": {
                        "description": "Redirected to frontend with desired redirect URL path",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Failed to login",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        },
...
        "/registrant": {
            "get": {
                "description": "Get registrants by userID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registrant"
                ],
                "summary": "Get by userID",
                "responses": {
                    "200": {
                        "description": "Failed to get registrants by user id",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-array_dto_RegistrantGetByUserID"
                        }
                    },
                    "500": {
                        "description": "Failed to get registrants by user id",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        },
        "/registrant/gen/{ticketID}": {
            "get": {
                "description": "Generate unique id for private mode",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registrant"
                ],
                "summary": "Generate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID",
                        "name": "ticketID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully generating unique id",
                        "schema": {
                            "type": ""
                        }
                    },
                    "500": {
                        "description": "Failed generating unique id",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/registrant/verify/{id}/{ticketID}": {
            "get": {
                "description": "Verify registrant by current user id and ticket id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registrant"
                ],
                "summary": "Verify",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID",
                        "name": "ticketID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Registrant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "303": {
                        "description": "Redirected to /verify?err=0?ticket=",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/registrant/{ticketID}": {
            "get": {
                "description": "Get registrants by ticketID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registrant"
                ],
                "summary": "Get by ticketID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID",
                        "name": "ticketID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "all",
                        "name": "all",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully get registrants by ticket id",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-dto_PaginatedRegistrants"
                        }
                    },
                    "500": {
                        "description": "Failed to get registrants by ticket id",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            },
            "post": {
                "description": "Create new registrant",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "registrant"
                ],
                "summary": "Create",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ticket ID",
                        "name": "ticketID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "state",
                        "name": "state",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully create registrant",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    },
                    "500": {
                        "description": "Failed to create registrant",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.APIResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/dto.ApiMeta"
                }
            }
        },
        "dto.APIResponse-array_dto_RegistrantGetByUserID": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RegistrantGetByUserID"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/dto.ApiMeta"
                }
            }
        },
...
        "dto.APIResponse-dto_PaginatedRegistrants": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.PaginatedRegistrants"
                },
                "meta": {
                    "$ref": "#/definitions/dto.ApiMeta"
                }
            }
        },
...
        "dto.ApiMeta": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.PaginatedRegistrants": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.RegistrantGetByTicketID"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "dto.RegistrantGetByTicketID": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "picture": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.RegistrantGetByUserID": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "ticket_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    }
}